refactor: filtering
